Extract config import tests
